+++ Docker/docker-compose-generate-env.sh
@@ -6,7 +6,6 @@ ADAGUC_DATA_DIR=${HOME}/adaguc-docker/adaguc-data
 ADAGUC_AUTOWMS_DIR=${HOME}/adaguc-docker/adaguc-autowms
 ADAGUC_DATASET_DIR=${HOME}/adaguc-docker/adaguc-datasets
 ADAGUC_NUMPARALLELPROCESSES=4
-ADAGUC_AUTOREMOVENONMATCHINGFILES=FALSE
 
 usage() { echo "Usage: $0 -p <port number> -e <external adress> -a <autowmsdir> -d <dataset dir> -f <datadir> -t <num parallel processes>" 1>&2; exit 1; }
 
@@ -60,7 +59,6 @@ echo "ADAGUC_AUTOWMS_DIR=${ADAGUC_AUTOWMS_DIR=}" >> .env
 echo "ADAGUC_DATASET_DIR=${ADAGUC_DATASET_DIR}" >> .env
 echo "ADAGUC_NUMPARALLELPROCESSES=${ADAGUC_NUMPARALLELPROCESSES}" >> .env
 echo "ADAGUC_PORT=${ADAGUC_PORT}" >> .env
-echo "ADAGUC_AUTOREMOVENONMATCHINGFILES=${ADAGUC_AUTOREMOVENONMATCHINGFILES}" >> .env
 echo "EXTERNALADDRESS=${EXTERNALADDRESS}" >> .env
 echo "############### env file ###############"
 cat .env